LayoutTests/fast/html/body-offset-properties.html

<body style="position: relative; width: 200px; margin: 0 auto;">
<pre id="console">The body should have 0 for both offsetLeft and offsetTop.

</pre>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
var console = document.getElementById("console");
console.appendChild(document.createTextNode("body.offsetParent = " + document.body.offsetParent + "\n"));
console.appendChild(document.createTextNode("body.offsetLeft = " + document.body.offsetLeft + "\n"));
console.appendChild(document.createTextNode("body.offsetTop = " + document.body.offsetTop + "\n"));
</script>

Property changes on: LayoutTests/fast/html/body-offset-properties.html
___________________________________________________________________
Name: svn:mime-type
   + text/html
Name: svn:eol-style
   + native
