LayoutTests/fast/html/body-offset-properties-expected.txt

The body should have 0 for both offsetLeft and offsetTop.

body.offsetParent = null
body.offsetLeft = 0
body.offsetTop = 0


Property changes on: LayoutTests/fast/html/body-offset-properties-expected.txt
___________________________________________________________________
Name: svn:eol-style
   + native
